+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 
 ## Fixed
 
+- Fix Dungeon Rewards Shuffle not generating with some settings. 
 - Fix Ikana Wisp showing up even if you did not beat Stone Tower Temple.
 - Fix some names in spoiler having their game as a suffix when they're unique.
 - Fix Moon Crash sometimes causing the Clock UI to disappear.